Backend/requirements.txt
pandas
scikit-learn
requests
flask
streamlit
google-generativeai
